Add JSON serialization for ShadowRoot
