+++ .gitignore
@@ -136,3 +136,4 @@ bazel-testlogs
 javascript/node/selenium-webdriver/.vscode/settings.json
 
 dotnet-bin
+.metadata/
